# app/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Fisher Fans API
  description: API pour gérer les utilisateurs, bateaux, sorties de pêche, réservations et carnets de pêche.
  version: 1.0.0
servers:
  - url: https://api.fisherfans.com/
    description: Serveur de production Fisher Fans
  - url: https://localhost:3000/
    description: Serveur de développement Fisher Fans
paths:
  /v1/users:
    post:
      summary: Créer un nouvel utilisateur
      tags:
        - Users
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/users/{id}:
    get:
      summary: Obtenir les détails d'un utilisateur spécifique
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Modifier un utilisateur existant
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Supprimer un utilisateur
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/trips:
    post:
      summary: Créer une nouvelle sortie de pêche
      tags:
        - Trips
      requestBody:
        $ref: '#/components/requestBodies/Trip'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    get:  # Nouveau endpoint pour récupérer des sorties de pêche avec filtres
      summary: Obtenir une liste de sorties de pêche avec des filtres
      tags:
        - Trips
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [journalière, récurrente]
          description: Filtrer par type de sortie
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filtrer par date de sortie
      responses:
        '200':
          description: Liste de sorties de pêche
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
  /v1/trips/{id}:
    get:
      summary: Obtenir les détails d'une sortie de pêche
      tags:
        - Trips
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          $ref: '#/components/schemas/Trip'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Modifier une sortie de pêche
      tags:
        - Trips
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        $ref: '#/components/requestBodies/Trip'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Supprimer une sortie de pêche
      tags:
        - Trips
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    SuccessResponse:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    ErrorResponse:
      description: Requête incorrecte ou ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  requestBodies:
    User:
      description: Détails de l'utilisateur
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'
...
    Trip:
      description: Détails de la sortie de pêche
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Trip'
...
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID unique de l'utilisateur
...
    TripId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID unique de la sortie de pêche
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de l'utilisateur
        name:
          type: string
          description: Nom de l'utilisateur
        email:
          type: string
          description: Adresse e-mail de l'utilisateur
        role:
          type: string
          description: Rôle de l'utilisateur (particulier, professionnel)
          enum: [particulier, professionnel]
        boats:
          type: array
          items:
            $ref: '#/components/schemas/Boat'
    Boat:
      type: object
      properties:
        id:
          type: integer
          description: ID unique du bateau
        name:
          type: string
          description: Nom du bateau
        boat_type:
          type: string
          description: Type du bateau
          enum: [open, cabin, catamaran, yacht, jetski, canoe]
        motor:
          type: string
          description: Type de motorisation
          enum: [diesel, essence, none]
        capacity:
          type: integer
          description: Capacité du bateau
        owner:
          $ref: '#/components/schemas/User'
    Trip:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de la sortie de pêche
        title:
          type: string
          description: Titre de la sortie de pêche
        trip_date:
          type: string
          format: date
          description: Date de la sortie
        price:
          type: number
          format: float
          description: Prix de la sortie de pêche
        boat:
          $ref: '#/components/schemas/Boat'
        organizer:
          $ref: '#/components/schemas/User'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicateur de succès de l'opération
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Message d'erreur
...